+++ src/jsdoc-globals.js
@@ -131,6 +131,7 @@ var TreePatch = {};
  * @property {Boolean} keyboard Support keyboard navigation (default: true).
  * @property {String} keyPathSeparator (default: "/")
  * @property {Integer} minExpandLevel 1: root node is not collapsible (default: 1)
+ * @property {jQuery} scrollParent: scrollable container for node.scrollIntiView() (default: $container)
  * @property {Integer} selectMode 1:single, 2:multi, 3:multi-hier (default: 2)
  * @property {any} source Used to Initialize the tree.
  * @property {object} strings Translation table